fix: rename app to all packagename and structure
