+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.wdionativedemoapp",
+    package = "com.wdiodemoapp",
 )
 
 android_resource(
     name = "res",
-    package = "com.wdionativedemoapp",
+    package = "com.wdiodemoapp",
     res = "src/main/res",
 )
