+++ android/app/build.gradle
@@ -144,7 +144,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.wdionativedemoapp"
+        applicationId "com.wdiodemoapp"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
