+++ android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.wdionativedemoapp">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.wdiodemoapp">
   <uses-permission android:name="android.permission.INTERNET"/>
   <application android:name=".MainApplication" android:label="@string/app_name" android:icon="@mipmap/ic_launcher" android:roundIcon="@mipmap/ic_launcher_round" android:allowBackup="false" android:theme="@style/AppTheme">
     <activity android:name=".SplashActivity" android:theme="@style/SplashTheme" android:label="@string/app_name">
