+++ android/app/src/main/java/com/wdiodemoapp/MainApplication.java
@@ -1,4 +1,4 @@
-package com.wdionativedemoapp;
+package com.wdiodemoapp;
 
 import android.app.Application;
 import android.content.Context;
@@ -69,7 +69,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.wdionativedemoapp.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.wdiodemoapp.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
