+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'wdioNativeDemoApp'
+rootProject.name = 'wdioDemoApp'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 // For the splashscreen
 include ':react-native-splash-screen'
