+++ src/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "wdioNativeDemoApp",
-  "displayName": "wdioNativeDemoApp"
+  "name": "wdioDemoApp",
+  "displayName": "wdioDemoApp"
 }
